@@ -2,7 +2,7 @@ using FluentAssertions;
 using NUnit.Framework;
 using System;
 using System.Data;
-using System.Data.SqlClient;
+using Microsoft.Data.SqlClient;
 using System.Threading.Tasks;
 using static CK.Testing.SqlServerTestHelper;
 
@@ -35,7 +35,7 @@ namespace CK.SqlServer.Tests
         }
 
         [Test]
-        public async Task SqlConnection_EnsureOpen_extension_asynchronous()
+        public async Task SqlConnection_EnsureOpen_extension_Async()
         {
             using( var oCon = new SqlConnection( TestHelper.GetConnectionString() ) )
             {
