@@ -15,9 +15,9 @@ namespace CodeCake.Abstractions
     {
         readonly StandardGlobalInfo _globalInfo;
         readonly string _typeName;
-        List<ArtifactFeed> _feeds;
-        List<ILocalArtifact> _artifacts;
-        List<ArtifactPush> _pushes;
+        List<ArtifactFeed>? _feeds;
+        List<ILocalArtifact>? _artifacts;
+        List<ArtifactPush>? _pushes;
 
         /// <summary>
         /// Initializes a new artifact type and adds it into
@@ -124,7 +124,7 @@ namespace CodeCake.Abstractions
         /// This uses the <see cref="GetPushListAsync(bool)"/> by default.
         /// </summary>
         /// <param name="pushes">Push details: defaults to the result of <see cref="GetPushListAsync"/>.</param>
-        public async Task PushAsync( IEnumerable<ArtifactPush> pushes = null )
+        public async Task PushAsync( IEnumerable<ArtifactPush>? pushes = null )
         {
             if( pushes == null ) pushes = await GetPushListAsync();
             var tasks = pushes.GroupBy( p => p.Feed ).Select( g => g.Key.PushAsync( g ) ).ToArray();
