@@ -1,6 +1,6 @@
 using Cake.Common.Diagnostics;
 using Cake.Core;
-using CK.Text;
+using CK.Core;
 using CodeCake.Abstractions;
 using CSemVer;
 using NuGet.Common;
@@ -101,9 +101,9 @@ namespace CodeCake
                     throw new NotSupportedException( "Should not be called in this scenario." );
                 }
 
-                PackageSource IPackageSourceProvider.GetPackageSourceByName( string name ) => _sources.Value.FirstOrDefault( s => s.Name == name );
+                PackageSource? IPackageSourceProvider.GetPackageSourceByName( string name ) => _sources.Value.FirstOrDefault( s => s.Name == name );
 
-                PackageSource IPackageSourceProvider.GetPackageSourceBySource( string source ) => _sources.Value.FirstOrDefault( s => s.Source == source );
+                PackageSource? IPackageSourceProvider.GetPackageSourceBySource( string source ) => _sources.Value.FirstOrDefault( s => s.Source == source );
 
                 void IPackageSourceProvider.RemovePackageSource( string name )
                 {
@@ -162,8 +162,8 @@ namespace CodeCake
                 public void LogError( string data ) { lock( _lock ) _ctx.Error( $"NuGet: {data}" ); }
                 public void LogSummary( string data ) { lock( _lock ) _ctx.Information( $"NuGet: {data}" ); }
                 public void LogInformationSummary( string data ) { lock( _lock ) _ctx.Information( $"NuGet: {data}" ); }
-                public void Log( LogLevel level, string data ) { lock( _lock ) _ctx.Information( $"NuGet ({level}): {data}" ); }
-                public Task LogAsync( LogLevel level, string data )
+                public void Log( NuGet.Common.LogLevel level, string data ) { lock( _lock ) _ctx.Information( $"NuGet ({level}): {data}" ); }
+                public Task LogAsync( NuGet.Common.LogLevel level, string data )
                 {
                     Log( level, data );
                     return System.Threading.Tasks.Task.CompletedTask;
@@ -205,7 +205,7 @@ namespace CodeCake
                     _ctx = ctx;
                 }
 
-                public string Id { get; }
+                public string Id { get; } = "Cake";
 
                 public Task<CredentialResponse> GetAsync(
                     Uri uri,
@@ -361,23 +361,21 @@ namespace CodeCake
                     Cake.Information( $"Pushing packages to '{Name}' => '{Url}'." );
                     var logger = InitializeAndGetLogger( Cake );
                     var updater = await _updater;
-                    foreach( var p in pushes )
-                    {
-                        string packageString = p.Name + "." + p.Version.WithBuildMetaData( null ).ToNormalizedString();
-                        var fullPath = ArtifactType.GlobalInfo.ReleasesFolder.AppendPart( packageString + ".nupkg" );
-                        await updater.Push(
-                            fullPath,
-                            string.Empty, // no Symbol source.
-                            20, //20 seconds timeout
-                            disableBuffering: false,
-                            getApiKey: endpoint => apiKey,
-                            getSymbolApiKey: symbolsEndpoint => null,
-                            noServiceEndpoint: false,
-                            skipDuplicate: true,
-                            symbolPackageUpdateResource: null,
-                            log: logger );
-                    }
-                    await OnAllArtifactsPushed( pushes );
+                    var names = pushes.Select( p => p.Name + "." + p.Version.WithBuildMetaData( null ).ToNormalizedString() );
+                    var fullPaths = names.Select( n => ArtifactType.GlobalInfo.ReleasesFolder.AppendPart( n + ".nupkg" ).ToString() );
+
+                    await updater.Push(
+                        fullPaths.ToList(),
+                        string.Empty, // no Symbol source.
+                        20, //20 seconds timeout
+                        disableBuffering: false,
+                        getApiKey: endpoint => apiKey,
+                        getSymbolApiKey: symbolsEndpoint => null,
+                        noServiceEndpoint: false,
+                        skipDuplicate: true,
+                        symbolPackageUpdateResource: null,
+                        log: logger );
+                    await OnAllArtifactsPushedAsync( pushes );
                 }
 
                 /// <summary>
@@ -386,7 +384,7 @@ namespace CodeCake
                 /// </summary>
                 /// <param name="pushes">The instances to push (that necessary target this feed).</param>
                 /// <returns>The awaitable.</returns>
-                protected virtual Task OnAllArtifactsPushed( IEnumerable<ArtifactPush> pushes )
+                protected virtual Task OnAllArtifactsPushedAsync( IEnumerable<ArtifactPush> pushes )
                 {
                     return System.Threading.Tasks.Task.CompletedTask;
                 }
@@ -395,7 +393,7 @@ namespace CodeCake
                 /// Must resolve the API key required to push the package.
                 /// </summary>
                 /// <returns>The secret (that can be null or empty).</returns>
-                protected abstract string ResolveAPIKey();
+                protected abstract string? ResolveAPIKey();
             }
         }
 
@@ -408,7 +406,7 @@ namespace CodeCake
         /// </summary>
         class VSTSFeed : NuGetHelper.NuGetFeed
         {
-            string _azureFeedPAT;
+            string? _azureFeedPAT;
 
             /// <summary>
             /// Initialize a new remote VSTS feed.
@@ -426,7 +424,7 @@ namespace CodeCake
             /// Gets the name of the environment variable that must contain the
             /// Personal Access Token that allows push to this feed.
             /// The  https://github.com/Microsoft/artifacts-credprovider VSS_NUGET_EXTERNAL_FEED_ENDPOINTS
-            /// will be dynalically generated.
+            /// will be dynamically generated.
             /// </summary>
             public override string SecretKeyName { get; }
 
@@ -436,7 +434,7 @@ namespace CodeCake
             /// </summary>
             /// <param name="ctx">The Cake context.</param>
             /// <returns>The "VSTS" API key or null to skip the push.</returns>
-            protected override string ResolveAPIKey()
+            protected override string? ResolveAPIKey()
             {
                 _azureFeedPAT = Cake.InteractiveEnvironmentVariable( SecretKeyName );
                 if( string.IsNullOrWhiteSpace( _azureFeedPAT ) )
@@ -476,10 +474,10 @@ namespace CodeCake
             /// </summary>
             /// <param name="organization">Name of the organization.</param>
             /// <param name="feedName">Identifier of the feed in Azure, inside the organization.</param>
-            public SignatureVSTSFeed( NuGetArtifactType t, string organization, string feedName, string projectName )
+            public SignatureVSTSFeed( NuGetArtifactType t, string organization, string feedName, string? projectName )
                 : base( t, organization + "-" + feedName,
-                      projectName != null ?
-                          $"https://pkgs.dev.azure.com/{organization}/{projectName}/_packaging/{feedName}/nuget/v3/index.json"
+                      projectName != null
+                        ? $"https://pkgs.dev.azure.com/{organization}/{projectName}/_packaging/{feedName}/nuget/v3/index.json"
                         : $"https://pkgs.dev.azure.com/{organization}/_packaging/{feedName}/nuget/v3/index.json",
                         GetSecretKeyName( organization ) )
             {
@@ -498,7 +496,7 @@ namespace CodeCake
             /// </summary>
             public string FeedName { get; }
 
-            public string ProjectName { get; }
+            public string? ProjectName { get; }
 
             /// <summary>
             /// Implements Package promotion in @CI, @Exploratory, @Preview, @Latest and @Stable views.
@@ -506,7 +504,7 @@ namespace CodeCake
             /// <param name="ctx">The Cake context.</param>
             /// <param name="pushes">The set of artifacts to promote.</param>
             /// <returns>The awaitable.</returns>
-            protected override async Task OnAllArtifactsPushed( IEnumerable<ArtifactPush> pushes )
+            protected override async Task OnAllArtifactsPushedAsync( IEnumerable<ArtifactPush> pushes )
             {
                 var basicAuth = Convert.ToBase64String( Encoding.ASCII.GetBytes( ":" + Cake.InteractiveEnvironmentVariable( SecretKeyName ) ) );
                 foreach( var p in pushes )
@@ -539,7 +537,7 @@ namespace CodeCake
                 }
             }
 
-            string GetPromotionJSONBody( string packageName, string packageVersion, string viewId, bool npm = false )
+            static string GetPromotionJSONBody( string packageName, string packageVersion, string viewId, bool npm = false )
             {
                 var bodyFormat = @"{
  ""data"": {
@@ -590,7 +588,7 @@ namespace CodeCake
             /// </summary>
             /// <param name="ctx">The Cake context.</param>
             /// <returns>The API key or null.</returns>
-            protected override string ResolveAPIKey()
+            protected override string? ResolveAPIKey()
             {
                 if( String.IsNullOrEmpty( SecretKeyName ) )
                 {
@@ -613,7 +611,7 @@ namespace CodeCake
 
             public override string SecretKeyName => null;
 
-            protected override string ResolveAPIKey() => null;
+            protected override string? ResolveAPIKey() => null;
         }
     }
 }
