@@ -1,13 +1,13 @@
 using Cake.Common.Diagnostics;
 using Cake.Common.IO;
 using Cake.Common.Solution;
-using Cake.Common.Tools.DotNetCore;
+using Cake.Common.Tools.DotNet;
 using Cake.Common.Tools.DotNetCore.Build;
 using Cake.Common.Tools.DotNetCore.Test;
 using Cake.Common.Tools.NUnit;
 using Cake.Core;
 using Cake.Core.IO;
-using CK.Text;
+using CK.Core;
 using CodeCake.Abstractions;
 using SimpleGitVersion;
 using System;
@@ -41,11 +41,10 @@ namespace CodeCake
         public readonly IReadOnlyList<SolutionProject> Projects;
         public readonly IReadOnlyList<SolutionProject> ProjectsToPublish;
 
-        DotnetSolution(
-            StandardGlobalInfo globalInfo,
-            string solutionFileName,
-            IReadOnlyList<SolutionProject> projects,
-            IReadOnlyList<SolutionProject> projectsToPublish )
+        DotnetSolution( StandardGlobalInfo globalInfo,
+                        string solutionFileName,
+                        IReadOnlyList<SolutionProject> projects,
+                        IReadOnlyList<SolutionProject> projectsToPublish )
         {
             _globalInfo = globalInfo;
             SolutionFileName = solutionFileName;
@@ -61,15 +60,15 @@ namespace CodeCake
         /// The predicate used to filter the project to publish. By default the predicate is p => !p.Path.Segments.Contains
         /// </param>
         /// <returns></returns>
-        public static DotnetSolution FromSolutionInCurrentWorkingDirectory(
-            StandardGlobalInfo globalInfo )
+        public static DotnetSolution FromSolutionInCurrentWorkingDirectory( StandardGlobalInfo globalInfo )
         {
             string solutionFileName = System.IO.Path.GetFileName(
                 globalInfo.Cake.GetFiles( "*.sln",
-                    new GlobberSettings
-                    {
-                        Predicate = p => !System.IO.Path.GetFileName( p.Path.FullPath ).EndsWith( ".local.sln", StringComparison.OrdinalIgnoreCase )
-                    } ).Single().FullPath
+                                          new GlobberSettings
+                                          {
+                                              Predicate = p => !System.IO.Path.GetFileName( p.Path.FullPath )
+                                                                              .EndsWith( ".local.sln", StringComparison.OrdinalIgnoreCase )
+                                          } ).Single().FullPath
             );
             var sln = globalInfo.Cake.ParseSolution( solutionFileName );
 
@@ -80,7 +79,7 @@ namespace CodeCake
                 .ToList();
             var projectsToPublish = projects.Where(
                     p => ((bool?)XDocument.Load( p.Path.FullPath )
-                        .Root
+                        .Root!
                         .Elements( "PropertyGroup" )
                         .Elements( "IsPackable" ).LastOrDefault() ?? true) == true
                 )
@@ -110,7 +109,7 @@ namespace CodeCake
             {
                 var exclude = new List<string>( excludedProjectsName ) { "CodeCakeBuilder" };
                 tempSln.ExcludeProjectsFromBuild( exclude.ToArray() );
-                _globalInfo.Cake.DotNetCoreBuild( tempSln.FullPath.FullPath,
+                _globalInfo.Cake.DotNetBuild( tempSln.FullPath.FullPath,
                     new DotNetCoreBuildSettings().AddVersionArguments( _globalInfo.BuildInfo, s =>
                     {
                         s.Configuration = _globalInfo.BuildInfo.BuildConfiguration;
@@ -118,7 +117,7 @@ namespace CodeCake
             }
         }
 
-        public void Test( IEnumerable<SolutionProject> testProjects = null )
+        public void Test( IEnumerable<SolutionProject>? testProjects = null )
         {
             if( testProjects == null )
             {
@@ -162,7 +161,7 @@ namespace CodeCake
                             _globalInfo.Cake.Information( $"Testing via NUnitLite ({framework}): {testBinariesPath}" );
                             if( !_globalInfo.CheckCommitMemoryKey( testBinariesPath ) )
                             {
-                                _globalInfo.Cake.DotNetCoreExecute( testBinariesPath );
+                                _globalInfo.Cake.DotNetExecute( testBinariesPath );
                             }
                         }
                     }
@@ -179,13 +178,13 @@ namespace CodeCake
                                 Framework = framework,
                                 NoRestore = true,
                                 NoBuild = true,
-                                Logger = "trx"
+                                Loggers = new List<string>() { "trx" }
                             };
                             if( _globalInfo.Cake.Environment.GetEnvironmentVariable( "DisableNodeReUse" ) != null )
                             {
                                 options.ArgumentCustomization = args => args.Append( "/nodeReuse:false" );
                             }
-                            _globalInfo.Cake.DotNetCoreTest( projectPath, options );
+                            _globalInfo.Cake.DotNetTest( projectPath, options );
                         }
                     }
 
